Move commit details retrieval out of `util`
