--- crates/util/src/git_author.rs
@@ -1,5 +1,0 @@
-/// Represents the common denominator of most git hosting authors
-#[derive(Debug)]
-pub struct GitAuthor {
-    pub avatar_url: String,
-}
