+++ crates/util/src/github.rs
@@ -1,7 +1,6 @@
-use crate::{git_author::GitAuthor, http::HttpClient};
+use crate::http::HttpClient;
 use anyhow::{anyhow, bail, Context, Result};
 use futures::AsyncReadExt;
-use isahc::{config::Configurable, AsyncBody, Request};
 use serde::Deserialize;
 use std::sync::Arc;
 use url::Url;
@@ -25,75 +24,6 @@ pub struct GithubRelease {
 pub struct GithubReleaseAsset {
     pub name: String,
     pub browser_download_url: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct CommitDetails {
-    commit: Commit,
-    author: Option<User>,
-}
-
-#[derive(Debug, Deserialize)]
-struct Commit {
-    author: Author,
-}
-
-#[derive(Debug, Deserialize)]
-struct Author {
-    email: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct User {
-    pub id: u64,
-    pub avatar_url: String,
-}
-
-pub async fn fetch_github_commit_author(
-    repo_owner: &str,
-    repo: &str,
-    commit: &str,
-    client: &Arc<dyn HttpClient>,
-) -> Result<Option<GitAuthor>> {
-    let url = format!("https://api.github.com/repos/{repo_owner}/{repo}/commits/{commit}");
-
-    let mut request = Request::get(&url)
-        .redirect_policy(isahc::config::RedirectPolicy::Follow)
-        .header("Content-Type", "application/json");
-
-    if let Ok(github_token) = std::env::var("GITHUB_TOKEN") {
-        request = request.header("Authorization", format!("Bearer {}", github_token));
-    }
-
-    let mut response = client
-        .send(request.body(AsyncBody::default())?)
-        .await
-        .with_context(|| format!("error fetching GitHub commit details at {:?}", url))?;
-
-    let mut body = Vec::new();
-    response.body_mut().read_to_end(&mut body).await?;
-
-    if response.status().is_client_error() {
-        let text = String::from_utf8_lossy(body.as_slice());
-        bail!(
-            "status error {}, response: {text:?}",
-            response.status().as_u16()
-        );
-    }
-
-    let body_str = std::str::from_utf8(&body)?;
-
-    serde_json::from_str::<CommitDetails>(body_str)
-        .map(|github_commit| {
-            if let Some(author) = github_commit.author {
-                Some(GitAuthor {
-                    avatar_url: author.avatar_url,
-                })
-            } else {
-                None
-            }
-        })
-        .context("deserializing GitHub commit details failed")
 }
 
 pub async fn latest_github_release(
